chore: Upgrade to Svelte 4
