+++ .github/workflows/index.yml
@@ -1,23 +1,24 @@
-name: "deploy"
+name: 'deploy'
 on:
-  push:
-    branches:
-      - main
+    push:
+        branches:
+            - main
+            - beta
 jobs:
-  release:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: '14'
-      - run: yarn install
-      - run: yarn test:ci
-      - run: yarn build
-      - run: npx semantic-release
-        env:
-          GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-      - uses: codecov/codecov-action@v1
-        with:
-          file: coverage/lcov.info
+    release:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v3
+            - uses: actions/setup-node@v3
+              with:
+                  node-version: 18
+            - run: yarn install
+            - run: yarn test:ci
+            - run: yarn build
+            - run: npx semantic-release
+              env:
+                  GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
+                  NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
+            - uses: codecov/codecov-action@v3
+              with:
+                  directory: ./coverage/
