+++ .gitignore
@@ -1,25 +1,28 @@
 # Build
-.cache/
-dist/
-stats.html
+/build
+/dist
+/package
+/.svelte-kit
 
-# Dependencies
-node_modules/
+# Env
+.env
+.env.*
 
 # Misc
 .DS_STORE
 
-# NPM
+# Dependencies
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 package-lock.json
 .yarn
+/node_modules
 
 # Webstorm
-.idea/
+/.idea
 *.iml
 
-# Jest
-coverage/
-__snapshots__/
+# Tests
+/coverage
+/__snapshots__
