+++ .prettierrc
@@ -1,29 +1,12 @@
 {
-  "printWidth": 120,
-  "useTabs": true,
-  "tabWidth": 4,
-  "singleQuote": true,
-  "semi": false,
-  "trailingComma": "es5",
-  "bracketSpacing": true,
-  "overrides": [
-    {
-      "files": "./**/*.js",
-      "options": {
-        "parser": "babylon"
-      }
-    },
-    {
-      "files": "./**/*.json",
-      "options": {
-        "parser": "json"
-      }
-    },
-    {
-      "files": ".prettierrc",
-      "options": {
-        "parser": "json"
-      }
-    }
-  ]
+	"printWidth": 120,
+	"useTabs": true,
+	"tabWidth": 4,
+	"singleQuote": true,
+	"semi": false,
+	"trailingComma": "es5",
+	"bracketSpacing": true,
+	"bracketSameLine": false,
+	"plugins": ["prettier-plugin-svelte"],
+	"overrides": [{ "files": "*.svelte", "options": { "parser": "svelte" } }]
 }
