+++ CHANGELOG.md
@@ -4,64 +4,56 @@
 
 # [1.5.0](https://github.com/untemps/svelte-readotron/compare/v1.4.4...v1.5.0) (2022-02-13)
 
-
 ### Features
 
-* Allow passing function in template prop ([#43](https://github.com/untemps/svelte-readotron/issues/43)) ([7334bc7](https://github.com/untemps/svelte-readotron/commit/7334bc7564b6bbb1ac03c1e0013158fd95d70e6f))
+-   Allow passing function in template prop ([#43](https://github.com/untemps/svelte-readotron/issues/43)) ([7334bc7](https://github.com/untemps/svelte-readotron/commit/7334bc7564b6bbb1ac03c1e0013158fd95d70e6f))
 
 ## [1.4.4](https://github.com/untemps/svelte-readotron/compare/v1.4.3...v1.4.4) (2022-02-11)
 
 ## [1.4.3](https://github.com/untemps/svelte-readotron/compare/v1.4.2...v1.4.3) (2022-01-14)
 
-
 ### Bug Fixes
 
-* Fix timeout value ([#40](https://github.com/untemps/svelte-readotron/issues/40)) ([0a6ee30](https://github.com/untemps/svelte-readotron/commit/0a6ee30b80fdc8e8a10df0febbc9cd876a585eeb))
+-   Fix timeout value ([#40](https://github.com/untemps/svelte-readotron/issues/40)) ([0a6ee30](https://github.com/untemps/svelte-readotron/commit/0a6ee30b80fdc8e8a10df0febbc9cd876a585eeb))
 
 ## [1.4.2](https://github.com/untemps/svelte-readotron/compare/v1.4.1...v1.4.2) (2022-01-13)
 
 ## [1.4.1](https://github.com/untemps/svelte-readotron/compare/v1.4.0...v1.4.1) (2022-01-13)
 
-
 ### Bug Fixes
 
-* Fix typo in import ([#38](https://github.com/untemps/svelte-readotron/issues/38)) ([937593c](https://github.com/untemps/svelte-readotron/commit/937593c1a51d03cd900b25593c964ee2e5b0ae75))
-* Remove duplicate file and fix wrong import ([#37](https://github.com/untemps/svelte-readotron/issues/37)) ([fb6f00b](https://github.com/untemps/svelte-readotron/commit/fb6f00b4489e7381e1a5f9a33f04f3756407a343))
+-   Fix typo in import ([#38](https://github.com/untemps/svelte-readotron/issues/38)) ([937593c](https://github.com/untemps/svelte-readotron/commit/937593c1a51d03cd900b25593c964ee2e5b0ae75))
+-   Remove duplicate file and fix wrong import ([#37](https://github.com/untemps/svelte-readotron/issues/37)) ([fb6f00b](https://github.com/untemps/svelte-readotron/commit/fb6f00b4489e7381e1a5f9a33f04f3756407a343))
 
 # [1.4.0](https://github.com/untemps/svelte-readotron/compare/v1.3.0...v1.4.0) (2021-02-03)
 
-
 ### Features
 
-* Add error slot ([50fc885](https://github.com/untemps/svelte-readotron/commit/50fc88593aa56239ffec5b96da4f1220ff2b9263))
+-   Add error slot ([50fc885](https://github.com/untemps/svelte-readotron/commit/50fc88593aa56239ffec5b96da4f1220ff2b9263))
 
 # [1.3.0](https://github.com/untemps/svelte-readotron/compare/v1.2.1...v1.3.0) (2021-01-30)
 
-
 ### Bug Fixes
 
-* Remove optional chaining syntax ([4fc302f](https://github.com/untemps/svelte-readotron/commit/4fc302fad101d279c3e4adeed7f87258ce71507d))
-
+-   Remove optional chaining syntax ([4fc302f](https://github.com/untemps/svelte-readotron/commit/4fc302fad101d279c3e4adeed7f87258ce71507d))
 
 ### Features
 
-* Add support for scroll progression ([8b3b8ca](https://github.com/untemps/svelte-readotron/commit/8b3b8ca7c902eaf3493ce789d83ebba8d30bbf50))
-* Change event name from 'scroll' to 'change' ([1a24456](https://github.com/untemps/svelte-readotron/commit/1a244560dfc3548286eb65c2117c68e7460aa9e3))
-* Dispatch scroll event ([0e69e3e](https://github.com/untemps/svelte-readotron/commit/0e69e3ec69716988393cefe36666f9fe584dbb03))
+-   Add support for scroll progression ([8b3b8ca](https://github.com/untemps/svelte-readotron/commit/8b3b8ca7c902eaf3493ce789d83ebba8d30bbf50))
+-   Change event name from 'scroll' to 'change' ([1a24456](https://github.com/untemps/svelte-readotron/commit/1a244560dfc3548286eb65c2117c68e7460aa9e3))
+-   Dispatch scroll event ([0e69e3e](https://github.com/untemps/svelte-readotron/commit/0e69e3ec69716988393cefe36666f9fe584dbb03))
 
 ## [1.2.1](https://github.com/untemps/svelte-readotron/compare/v1.2.0...v1.2.1) (2021-01-22)
 
 # [1.2.0](https://github.com/untemps/svelte-readotron/compare/v1.1.3...v1.2.0) (2021-01-20)
 
-
 ### Bug Fixes
 
-* Fix wrong import in tests ([#11](https://github.com/untemps/svelte-readotron/issues/11)) ([a17bb24](https://github.com/untemps/svelte-readotron/commit/a17bb246fef75704f4dc230d88756526b962c703))
-
+-   Fix wrong import in tests ([#11](https://github.com/untemps/svelte-readotron/issues/11)) ([a17bb24](https://github.com/untemps/svelte-readotron/commit/a17bb246fef75704f4dc230d88756526b962c703))
 
 ### Features
 
-* Add words token ([7e37756](https://github.com/untemps/svelte-readotron/commit/7e377562e149957ac700d6f8bc34c07dadf6fbe7))
+-   Add words token ([7e37756](https://github.com/untemps/svelte-readotron/commit/7e377562e149957ac700d6f8bc34c07dadf6fbe7))
 
 ## [1.1.3](https://github.com/untemps/svelte-readotron/compare/v1.1.2...v1.1.3) (2021-01-19)
 
@@ -71,16 +63,14 @@
 
 # [1.1.0](https://github.com/untemps/svelte-readotron/compare/v1.0.1...v1.1.0) (2021-01-11)
 
-
 ### Features
 
-* Add slot support ([#4](https://github.com/untemps/svelte-readotron/issues/4)) ([02f1500](https://github.com/untemps/svelte-readotron/commit/02f1500ca2d71bf77389812c26ce3a35e3b375e6))
+-   Add slot support ([#4](https://github.com/untemps/svelte-readotron/issues/4)) ([02f1500](https://github.com/untemps/svelte-readotron/commit/02f1500ca2d71bf77389812c26ce3a35e3b375e6))
 
 ## [1.0.1](https://github.com/untemps/svelte-readotron/compare/v1.0.0...v1.0.1) (2021-01-06)
 
 # 1.0.0 (2021-01-05)
 
-
 ### Features
 
-* Initiate component ([3c1f907](https://github.com/untemps/svelte-readotron/commit/3c1f907180d0998be37a6b8c49f7c999b0b837a1))
+-   Initiate component ([3c1f907](https://github.com/untemps/svelte-readotron/commit/3c1f907180d0998be37a6b8c49f7c999b0b837a1))
