--- babel.config.js
@@ -1,10 +1,0 @@
-module.exports = api => ({
-	"presets": [
-		"@babel/preset-env"
-	],
-	"plugins": [
-		"@babel/plugin-proposal-class-properties",
-		...(api.env('test') ? ["@babel/plugin-transform-runtime"] : [])
-	]
-})
-
