--- dev/public/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset='utf-8'>
-    <meta name='viewport' content='width=device-width,initial-scale=1'>
-    <title>svelte-readotron</title>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Vollkorn:ital,wght@400;600&display=swap">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Rubik:wght@400;600&display=swap">
-    <link href="https://css.gg/css?=|options" rel="stylesheet">
-    <link rel='stylesheet' href='./index.css'>
-    <style>
-        html, body {
-            width: 100%;
-            margin: 0;
-            padding: 0;
-            font-family: Rubik, sans-serif;
-            font-size: 14px;
-        }
-    </style>
-    <script src="./index.js" type="module"></script>
-</head>
-<body>
-</body>
-</html>
