+++ package.json
@@ -2,7 +2,9 @@
 	"name": "@untemps/svelte-readotron",
 	"version": "1.5.2",
 	"author": "Vincent Le Badezet <v.lebadezet@untemps.net>",
+	"homepage": "https://github.com/untemps/svelte-readotron",
 	"repository": "git@github.com:untemps/svelte-readotron.git",
+	"bugs": "https://github.com/untemps/svelte-readotron/issues",
 	"license": "MIT",
 	"description": "Svelte component to display an estimated reading time",
 	"keywords": [
@@ -20,78 +22,63 @@
 	"publishConfig": {
 		"access": "public"
 	},
-	"main": "dist/index.js",
-	"module": "dist/index.es.js",
-	"svelte": "dist/index.es.js",
+	"svelte": "./dist/index.js",
+	"types": "./dist/index.d.ts",
+	"type": "module",
 	"files": [
-		"dist"
+		"dist",
+		"!dist/**/*.test.*"
 	],
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"svelte": "./dist/index.js"
+		}
+	},
+	"engines": {
+		"node": ">=18"
+	},
 	"devDependencies": {
-		"@babel/cli": "^7.18.6",
-		"@babel/core": "^7.18.6",
-		"@babel/plugin-proposal-class-properties": "^7.18.6",
-		"@babel/plugin-transform-runtime": "^7.18.6",
-		"@babel/preset-env": "^7.18.6",
-		"@commitlint/cli": "^17.0.3",
-		"@commitlint/config-conventional": "^17.0.3",
-		"@rollup/plugin-babel": "^5.3.1",
-		"@rollup/plugin-commonjs": "^22.0.1",
-		"@rollup/plugin-node-resolve": "^13.3.0",
-		"@semantic-release/changelog": "^6.0.1",
+		"@commitlint/cli": "^17.7.2",
+		"@commitlint/config-conventional": "^17.7.0",
+		"@semantic-release/changelog": "^6.0.3",
 		"@semantic-release/git": "^10.0.1",
-		"@semantic-release/github": "^8.0.5",
-		"@testing-library/dom": "^8.16.0",
-		"@testing-library/jest-dom": "^5.16.4",
-		"@testing-library/svelte": "^3.1.3",
-		"babel-jest": "^28.1.3",
+		"@semantic-release/github": "^9.2.1",
+		"@sveltejs/adapter-auto": "^2.0.0",
+		"@sveltejs/kit": "^1.20.4",
+		"@sveltejs/package": "^2.0.0",
+		"@testing-library/jest-dom": "^6.1.3",
+		"@testing-library/svelte": "^4.0.3",
+		"@vitest/coverage-v8": "^0.34.6",
 		"cross-env": "^7.0.3",
-		"husky": "^8.0.1",
-		"jest": "^28.1.3",
-		"jest-environment-jsdom": "^28.1.3",
-		"prettier": "^2.7.1",
-		"rimraf": "^3.0.2",
-		"rollup": "^2.77.0",
-		"rollup-plugin-sizes": "^1.0.4",
-		"rollup-plugin-svelte": "^7.0.0",
-		"rollup-plugin-terser": "^7.0.2",
-		"rollup-plugin-visualizer": "^5.7.1",
-		"semantic-release": "^19.0.2",
-		"svelte": "^3.49.0",
-		"svelte-jester": "^2.3.2"
+		"husky": "^8.0.3",
+		"jsdom": "^22.1.0",
+		"prettier": "^3.0.3",
+		"prettier-plugin-svelte": "^3.0.3",
+		"publint": "^0.1.9",
+		"semantic-release": "^22.0.5",
+		"svelte": "^4.0.5",
+		"tslib": "^2.4.1",
+		"typescript": "^5.0.0",
+		"vite": "^4.4.2",
+		"vitest": "^0.34.0"
 	},
 	"dependencies": {
-		"@untemps/dom-observer": "^2.0.0",
-		"@untemps/read-per-minute": "^1.0.3",
-		"@untemps/utils": "^2.0.0"
+		"@untemps/read-per-minute": "^2.0.1",
+		"@untemps/svelte-use-drop-outside": "^1.6.0",
+		"@untemps/svelte-use-tooltip": "^2.8.0",
+		"@untemps/utils": "^3.0.0"
 	},
-	"jest": {
-		"transform": {
-			"^.+\\.js$": "babel-jest",
-			"^.+\\.svelte$": "svelte-jester"
-		},
-		"moduleFileExtensions": [
-			"js",
-			"svelte"
-		],
-		"coverageDirectory": "./coverage/",
-		"collectCoverage": true,
-		"coveragePathIgnorePatterns": [
-			"<rootDir>/src/.*/.*.svelte"
-		],
-		"setupFilesAfterEnv": [
-			"<rootDir>/jest/jest.setup.js"
-		],
-		"restoreMocks": true
-	},
-	"husky": {
-		"hooks": {
-			"pre-commit": "yarn test:ci && yarn prettier",
-			"commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-		}
+	"peerDependencies": {
+		"svelte": "^4.0.0"
 	},
 	"release": {
 		"branches": [
-			"main"
+			"main",
+			{
+				"name": "beta",
+				"prerelease": true
+			}
 		],
 		"plugins": [
 			[
@@ -99,7 +86,6 @@
 				{
 					"releaseRules": [
 						{
-							"type": "chore",
 							"scope": "force",
 							"release": "patch"
 						}
@@ -115,16 +101,7 @@
 				{
 					"assets": [
 						{
-							"path": "dist/index.js",
-							"label": "CJS distribution"
-						},
-						{
-							"path": "dist/index.es.js",
-							"label": "ES distribution"
-						},
-						{
-							"path": "dist/index.umd.js",
-							"label": "UMD distribution"
+							"path": "dist/index.js"
 						}
 					]
 				}
@@ -132,13 +109,15 @@
 		]
 	},
 	"scripts": {
-		"dev": "cd dev && rimraf dist && yarn && yarn start",
-		"test": "jest -u --watch",
-		"test:ci": "jest -u -b --ci --coverage",
-		"build": "rimraf dist && yarn build:cjs && yarn build:es && yarn build:umd",
-		"build:cjs": "cross-env NODE_ENV=production BABEL_ENV=cjs rollup -c",
-		"build:es": "cross-env NODE_ENV=production BABEL_ENV=es rollup -c",
-		"build:umd": "cross-env NODE_ENV=production BABEL_ENV=umd rollup -c",
-		"prettier": "prettier \"*/**/*.js\" --ignore-path ./.prettierignore --write && git add . && git status"
+		"dev": "vite dev",
+		"build": "vite build && yarn package",
+		"preview": "vite preview",
+		"package": "svelte-kit sync && svelte-package && publint",
+		"prepublishOnly": "yarn package",
+		"test": "vitest watch --coverage",
+		"test:ci": "vitest run --coverage",
+		"lint": "prettier --check .",
+		"format": "prettier --write .",
+		"prepare": "husky install"
 	}
 }
