+++ src/lib/components/__tests__/ErrorSlotTest.svelte
