+++ src/lib/scroll/ScrollProgress.js
