+++ src/routes/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import App from '../components/App.svelte'
+</script>
+
+<App />
