--- src/utils/isFunction.js
@@ -1,1 +1,0 @@
-export default (val) => typeof val === 'function'
