+++ static/app.css
@@ -1,0 +1,8 @@
+* {
+	box-sizing: border-box;
+}
+
+html,
+body {
+	height: 100%;
+}
