+++ vite.config.js
@@ -1,0 +1,6 @@
+import { sveltekit } from '@sveltejs/kit/vite'
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+	plugins: [sveltekit()],
+})
